requirements.txt
python-dotenv
binance-connector